+++ src/routes.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Route, IndexRoute } from 'react-router';
+import { Route, IndexRoute, Redirect } from 'react-router';
 
 import App from './containers/App';
 import Home from './containers/HomePage';
@@ -7,7 +7,6 @@ import About from './containers/AboutPage';
 import TagFilter from './containers/TagFilterPage';
 import ProjectPage from './containers/ProjectPage';
 import TextFilter from './containers/TextFilterPage';
-import loading from './loading';
 
 function getRoutes() {
 
@@ -23,6 +22,7 @@ function getRoutes() {
       <Route path="projects/:id" component={ProjectPage}/>
       <Route path="tags/:id" component={ TagFilter } />
       <Route path="search/:text" component={ TextFilter }/>
+      <Redirect from="*" to="home" />
     </Route>
   );
   return routes;