+++ test/components/ProjectSpec.js
@@ -5,6 +5,7 @@ import TestUtils from 'react-addons-test-utils';
 
 import {getAllTags, getTag, getAllComponents, getComponent} from '../utils';
 import setup from '../setup.js';
+import populate from '../../src/helpers/populate';
 
 
 //Data
@@ -17,17 +18,23 @@ import * as actions from '../../src/actions';
 setup();
 
 //Components to check
-import ProjectPage from '../../src/components/projects/ProjectPage';
+import Project from '../../src/components/projects/Project';
 
 const state = getInitialState(data);
-state.project = state.allProjects[0];
+const {
+  entities: { projects, tags },
+} = state;
 
+const id = 'react';
 
-test('Check ProjectPage component', (assert) => {
+let project = projects[id];
+project = populate(tags)(project);
+
+test('Check <Project> component', (assert) => {
 
   const component = TestUtils.renderIntoDocument(
-    <ProjectPage
-      githubProjects = { state }
+    <Project
+      project = { project }
       actions = { actions }
     />
   );
@@ -36,8 +43,8 @@ test('Check ProjectPage component', (assert) => {
 
   const h1 = getTag(component, 'h1');
   const h1Node = ReactDOM.findDOMNode(h1);
-  const title = state.project.name;
-  assert.equal(h1Node.innerHTML, title, `Project title '${title}' should be displayed`);
+  const title = project.name;
+  assert.equal(h1Node.innerHTML, title, `Project name '${title}' should be displayed`);
 
   assert.end();