+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## 0.2.5 - 2015-12-12
+* Redux optimization, use middleware for tracking and navigation side effects (scroll to top, close the side menu on mobiles), create `containers` connected to Redux state
+* Redirect to homepage for unknown URLs
+
 ## 0.2.4 - 2015-11-28
 * Get projects.json from firebase instead of divshot (issue #7)
 * Use the new logo in the splash screen